hurt, language, cleaner code
fixed a bug of the party member state not resetting upon finishing the fight action
char state changed to use indices instead of numbers for better readability
added hurt sprites for party in the encounter menu when getting hit
some saves will now be marked as "incompatible" if made in an old version of the engine that exported saves in some way different
loc_switch_lang now is not required to restart the game, rather just restarting the room.
full save select localization
rooms are now stored as localization ids
o_world user event 0 is now used as the event that is called every time the language is switched. useful for importing sprites and fonts that will be used throughout the game